+++ src/L10nServiceProvider.php
@@ -5,10 +5,12 @@ namespace Goodcat\L10n;
 use Goodcat\L10n\Listeners\RegisterLocalizedViewsPath;
 use Goodcat\L10n\Mixin\LocalizedApplication;
 use Goodcat\L10n\Mixin\LocalizedRoute;
+use Goodcat\L10n\Mixin\LocalizedRouter;
 use Goodcat\L10n\Routing\LocalizedUrlGenerator;
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Events\LocaleUpdated;
 use Illuminate\Routing\Route;
+use Illuminate\Routing\Router;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\ServiceProvider;
 
@@ -48,19 +50,9 @@ class L10nServiceProvider extends ServiceProvider
 
         $this->app->singleton(RegisterLocalizedViewsPath::class);
 
+        Router::mixin(new LocalizedRouter);
         Application::mixin(new LocalizedApplication);
         Route::mixin(new LocalizedRoute);
-
-        /*
-        | The RouteRegistrar::class doesn't implement
-        | the Macroable trait on Laravel v11.
-        | I'll wait for the next major release.
-       */
-
-//        $langMixin = new LangMixin;
-
-//        RouteRegistrar::mixin($langMixin);
-//        Router::mixin($langMixin);
     }
 
     protected function requestRebinder(): \Closure